2:49
User registration

+++ base/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
 from django.db.models import Q
 from django.contrib.auth.decorators import login_required 
@@ -10,6 +11,9 @@ from .forms import RoomForm
 # Create your views here.
 
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
     
     if request.method== 'POST':
         username = request.POST.get('username')
@@ -28,12 +32,16 @@ def loginPage(request):
         else:
             message.error(request, "username or password does not exist")
 
-    context={}
+    context={'page':page}
     return render(request, 'base/login_register.html', context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+def registerPage(request):
+    page= 'register'
+    return render(request, "base/login_register.html"),
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -70,6 +78,9 @@ def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
 
+    if request.user != room.host:
+        return HttpResponse(' You are not allowed here')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -79,8 +90,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request,'base/room_form.html', context)
 
+@login_required(login_url="login")
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse(' You are not allowed here')
+
     if request.method == 'POST':
         room.delete()
         return redirect("home")
